moved aggregation from network to geometry_operations created am example for aggregation translated comments
