+++ dhnx/network.py
@@ -25,6 +25,7 @@ from .optimization.optimization_models import optimize_operation
 from .optimization.optimization_models import setup_optimise_investment
 from .optimization.optimization_models import solve_optimisation_investment
 from .simulation import simulate
+from .gistools.geometry_operations import aggregation
 
 dir_name = os.path.dirname(__file__)
 
@@ -318,279 +319,12 @@ class ThermalNetwork():
     def simulate(self, *args, **kwargs):
         self.results.simulation = simulate(self, *args, **kwargs)
 
-    def aggregate(self, maxlength = 100):
-        """
-        :param maxlength:
-        :return:
-        """
+    def aggregate(self):
+
         self.aggregatednetwork = dict()
-        test_dict = aggregation(forks = self.components["forks"],
-                                pipes = self.components["pipes"],
-                                consumers = self.components["consumers"],
-                                producers = self.components["producers"])
-
-# TODO: Straßen werden als ganzes aggregiert, obwohl das nicht immer sinnvoll ist (siehe Straße "Im Grund" oben rechts)
-
-def aggregation(forks, pipes, consumers, producers):
-
-    from shapely import geometry, ops
-    import geopandas as gpd
-    import matplotlib.pyplot as plt
-    import pandas as pd
-
-# TODO: Warnings anschalten
-    import warnings
-    warnings.filterwarnings("ignore")
-
-    # # # Identifizierung der Superforks
-    # DL und GL aus pipes speichern
-    DLGLpipes = pipes.loc[pipes['type'].isin(['DL', 'GL'])]
-
-    # Anzahl von Verbindungen von den forks zählen (from_node und to_node)
-    count_forks_from_node = DLGLpipes['from_node'].value_counts()
-    count_forks_to_node = DLGLpipes['to_node'].value_counts()
-
-    # Anzahl von from_node und to_node addieren
-    count_forks = count_forks_from_node.add(count_forks_to_node, fill_value=0)
-
-    # Die Superforks, also forks mit 1, 3, 4 Verbindungen, identifizieren
-    count_superforks = count_forks[count_forks.isin([1, 3, 4])]
-
-    # Index von den Superforks speichern
-    superforks_indexlist = count_superforks.index.tolist()
-
-    # Superforks aus forks nehmen, wenn die forks als Superfork identifiziert wurden
-    super_forks = forks.copy()
-    super_forks = super_forks[super_forks['id_full'].isin(superforks_indexlist)]
-    super_forks = super_forks.reset_index(drop=True)
-
-    # # # 2. SuperPipes identifizieren und mergen
-
-    # super_pipes initialisieren
-    super_pipes = pipes.copy()
-    super_pipes = super_pipes.drop(range(0, len(pipes)))
-    # Liste für aggregierte forks und pipes initialisieren
-    aggregated_forks = []
-    aggregated_pipes = []
-
-    # HL pipes für aggregated_HLpipes_segment_i_a
-    HLpipes = pipes.loc[pipes['type'].isin(['HL'])]
-
-    # i = -1
-    i = -1  # # # Erste Schleife: Superforks durchgehen
-
-    while i <= len(super_forks) - 2:  # länge ist absolut und id startet bei 0
-        i += 1
-        # Den aktuellen super_fork (i) aus der super_fork Liste auswählen ...
-        superfork_i_id_full = super_forks.loc[i]['id_full']
-        # ... und zu den aggregierten hinzugügen
-        aggregated_forks.append(superfork_i_id_full)
-
-        # Consumer die mit dem super-fork verbunden sind Heraussuchen
-        aggregated_consumer_super_fork_i = []
-        aggregated_consumer_super_fork_i = HLpipes.loc[HLpipes['from_node'] == superfork_i_id_full]['to_node'].tolist()
-        str_aggregated_consumer_super_forks = ', '.join(aggregated_consumer_super_fork_i)
-        index = super_forks.loc[super_forks['id_full'] == superfork_i_id_full].index[0]
-        super_forks.at[index, 'aggregated_consumers'] = str_aggregated_consumer_super_forks
-
-        # max heat raus suchen
-        aggregated_P_heat_super_fork_i = []
-        aggregated_P_heat_super_fork_i = consumers.loc[consumers['id_full'].isin(aggregated_consumer_super_fork_i)][
-            'P_heat_max'].tolist()
-        sum_aggregated_P_heat_max_super_fork_i = sum(aggregated_P_heat_super_fork_i)
-        super_forks.at[index, 'aggregated_P_heat_max'] = sum_aggregated_P_heat_max_super_fork_i
-
-        # Pipes die am Superfork i verbunden sind raus suchen und in einen GeoDataFrame speichern
-        pipes_superfork_i_from_node = DLGLpipes[DLGLpipes['from_node'] == superfork_i_id_full]
-        pipes_superfork_i_to_node = DLGLpipes[DLGLpipes['to_node'] == superfork_i_id_full]
-        pipes_superfork_i = pipes_superfork_i_from_node.append(pipes_superfork_i_to_node)
-        pipes_superfork_i = pipes_superfork_i.reset_index(drop=True)
-
-        a = - 1  # # # Zweite Schleife: Superpipes a vom Superfork i durchgehen
-        while a <= len(pipes_superfork_i) - 2:  # länge ist absolut und id startet bei 0
-            a += 1
-            # Wenn die pipe schon aggregiert wurde, soll die nächst pipe überprüft werden
-            if pipes_superfork_i.loc[a]['id'] in aggregated_pipes:
-                continue  # evtl. a += 1 vorher break oder continue
-            # Ein Segment ist die Anreihung von pipes bevor diese gemerget werden
-            # Das Segment wird initialisiert und die pipe a ist die erste pipe des Segments
-            segment_i_a = pipes_superfork_i.copy()
-            segment_i_a = segment_i_a[segment_i_a.index == a]
-            segment_i_a = segment_i_a.reset_index(drop=True)
-            included_forks_segment_i_a = []
-
-            aggregation_segment = False  # # # Dritte Schleife: Die Elemente b von pipe a durchgehen
-            b = 0
-            while aggregation_segment == False:
-
-                # print('Superfork i: ', superfork_i_id_full)
-                # print('pipe a: ', segment_i_a.at[0, 'id'])
-                # print('segment b: ', b)
-
-                # Heraussuchen des nächsten forks der mit der pipe b verbunden ist
-                fork_from_pipe_b_segment_i_a = 0  # evtl nicht nötig
-                fork_to_pipe_b_segment_i_a = 0  # evtl nicht nötig
-                fork_from_pipe_b_segment_i_a = segment_i_a.at[b, 'from_node']  # hier tritt ein fehler auf
-                fork_to_pipe_b_segment_i_a = segment_i_a.at[b, 'to_node']
-                count_type_forks_pipe_b_segment_i_a = 0
-                status_fork_from_pipe_b_segment_i_a = 0  # nicht benötigt
-                status_fork_to_pipe_b_segment_i_a = 0  # nicht benötigt
-                # Initialisieren des nächsten forks
-                fork_next_segment_i_a = 0
-                # TODO: Kann der nächste Fork überhaupt in aggregated forks sein, wenn die pipe nicht aggregiert ist?
-                # Prüfen ob der fork 'from' bereits aggregiert oder ein superfork ist oder ein producer ist
-                # Wenn ja wird die Anzahl der count_type_forks_pipe_b_segment_i_a erhöht ...
-                # ... falls nicht ist der fork zum nächsten segment der fork 'from'
-                if fork_from_pipe_b_segment_i_a in super_forks[
-                    'id_full'].unique() or fork_from_pipe_b_segment_i_a in aggregated_forks or fork_from_pipe_b_segment_i_a in \
-                        producers['id_full'].unique():
-                    count_type_forks_pipe_b_segment_i_a += 1
-                    status_fork_from_pipe_b_segment_i_a = 'aggregated/superfork'  # nicht benötigt
-                else:
-                    fork_next_segment_i_a = fork_from_pipe_b_segment_i_a  # Der nächste fork darf noch nicht aggregiert sein
-                # Prüfen ob der fork 'to' bereits aggregiert oder ein superfork ist oder ein producer ist
-                # Wenn ja wird die Anzahl der count_type_forks_pipe_b_segment_i_a erhöht
-                # ... falls nicht ist der fork zum nächsten segment der fork 'to'
-                if fork_to_pipe_b_segment_i_a in super_forks[
-                    'id_full'].unique() or fork_to_pipe_b_segment_i_a in aggregated_forks or fork_to_pipe_b_segment_i_a in \
-                        producers['id_full'].unique():
-                    count_type_forks_pipe_b_segment_i_a += 1
-                    status_fork_to_pipe_b_segment_i_a = 'aggregated/superfork'  # nicht benötigt
-                else:
-                    fork_next_segment_i_a = fork_to_pipe_b_segment_i_a  # Der nächste fork darf noch nicht aggregiert sein
-                # Wenn beide forks aggregiert sind wird das Segment gemergt
-                if count_type_forks_pipe_b_segment_i_a == 2:
-                    # merge new geometry
-                    geom_multi_segment_i_a = geometry.MultiLineString(segment_i_a['geometry'].unique())
-                    geom_line_segment_i_a = ops.linemerge(geom_multi_segment_i_a)
-                    # create new pipe
-                    merged_segment_i_a = segment_i_a[segment_i_a.index == 0]
-                    merged_segment_i_a['geometry'] = geom_line_segment_i_a
-
-                    # # set from_node and to_node of the merged segment
-                    # from_node is the super fork i
-                    merged_segment_i_a.at[0, 'from_node'] = superfork_i_id_full
-                    # to_node is not the super fork i and ether a super fork or a producer
-                    last_fork_segment_i_a = 0
-                    if fork_to_pipe_b_segment_i_a != superfork_i_id_full and (fork_to_pipe_b_segment_i_a in super_forks[
-                        'id_full'].unique() or fork_to_pipe_b_segment_i_a in \
-                                                                              producers['id_full'].unique()):
-
-                        last_fork_segment_i_a = fork_to_pipe_b_segment_i_a
-
-                    elif fork_from_pipe_b_segment_i_a != superfork_i_id_full and (
-                            fork_from_pipe_b_segment_i_a in super_forks[
-                        'id_full'].unique() or fork_from_pipe_b_segment_i_a in \
-                            producers['id_full'].unique()):
-
-                        last_fork_segment_i_a = fork_from_pipe_b_segment_i_a
-
-                    else:
-                        print('error: last fork is not a')
-
-                    merged_segment_i_a.at[0, 'to_node'] = last_fork_segment_i_a
-                    # #
-                    # # add column 'included_forks'
-                    # merged_segment_i_a['aggregated_forks'] = merged_segment_i_a['aggregated_forks'].astype(object)
-                    str_included_forks_segment_i_a = ', '.join(included_forks_segment_i_a)
-                    merged_segment_i_a.at[0, 'included_forks'] = str_included_forks_segment_i_a
-
-                    # # add column 'aggregated_consumers'
-                    included_consumer_segment_i_a = []
-                    included_consumer_segment_i_a = HLpipes.loc[HLpipes['from_node'].isin(included_forks_segment_i_a)][
-                        'to_node'].tolist()
-                    str_included_consumer_segment_i_a = ', '.join(included_consumer_segment_i_a)
-                    merged_segment_i_a.at[0, 'included_consumer'] = str_included_consumer_segment_i_a
-
-                    # # add column 'aggregated_P_heat_max'
-                    included_P_heat_max_segment_i_a = []
-                    included_P_heat_max_segment_i_a = \
-                    consumers.loc[consumers['id_full'].isin(included_consumer_segment_i_a)]['P_heat_max'].tolist()
-                    sum_included_P_heat_max_segment_i_a = sum(included_P_heat_max_segment_i_a)
-                    merged_segment_i_a.at[0, 'included_P_heat_max'] = sum_included_P_heat_max_segment_i_a
-
-                    # # add column 'Gleichzeitigkeitsfaktor'
-                    number_included_consumer_segment_i_a = len(included_consumer_segment_i_a)
-                    Gleichzeitigkeitsfaktor_segment_i_a = pow(1.05, -number_included_consumer_segment_i_a)
-                    merged_segment_i_a.at[0, 'Gleichzeitigkeitsfaktor'] = Gleichzeitigkeitsfaktor_segment_i_a
-
-                    # add new pipe to super pipes
-                    super_pipes = super_pipes.append(merged_segment_i_a)
-                    # add pipe_ids to aggregated pipes
-                    aggregated_pipes = aggregated_pipes + segment_i_a['id'].tolist()
-
-                    aggregation_segment == True
-
-                    break
-
-                    # Identifizieren welche pipes mit fork_next_segment_i_a verbunden sind
-                elif count_type_forks_pipe_b_segment_i_a == 1:  #
-
-                    pipe_next_segment_i_a = 0  #
-                    # Nächste Pipe darf nicht die aktuelle sein!
-                    # TODO:Abfrage schlauer gestalten
-
-                    list_of_connected_pipes_to_next_fork = []
-                    list_of_connected_pipes_to_next_fork = list_of_connected_pipes_to_next_fork + DLGLpipes.loc[
-                        DLGLpipes['from_node'].isin([fork_next_segment_i_a])]['id'].tolist() + DLGLpipes.loc[
-                                                               DLGLpipes['to_node'].isin([fork_next_segment_i_a])][
-                                                               'id'].tolist()
-
-                    # print('list of connected pipes: ', list_of_connected_pipes_to_next_fork)
-
-                    if segment_i_a.at[b, 'id'] == list_of_connected_pipes_to_next_fork[0]:
-                        pipe_next_segment_i_a = DLGLpipes.loc[
-                            DLGLpipes['id'].isin([list_of_connected_pipes_to_next_fork[1]])]
-
-                    elif segment_i_a.at[b, 'id'] == list_of_connected_pipes_to_next_fork[1]:
-                        pipe_next_segment_i_a = DLGLpipes.loc[
-                            DLGLpipes['id'].isin([list_of_connected_pipes_to_next_fork[0]])]
-
-                    else:
-                        print('error: next fork doesnt connect to any new pipe')
-
-                    # b hoch zählen
-                    b += 1
-
-                    # hinzufügen zu segment i_a
-                    segment_i_a = segment_i_a.append(pipe_next_segment_i_a)
-
-                    # index resetten
-                    segment_i_a = segment_i_a.reset_index(drop=True)
-
-                    # fork zu aggregated fork
-                    aggregated_forks.append(fork_next_segment_i_a)
-
-                    # aggregated forks segment i a
-                    included_forks_segment_i_a.append(fork_next_segment_i_a)
-                else:
-                    print('error: pipe is not connected to any aggregated fork or super fork')
-
-    # Länge der super_pipes berechnen
-    super_pipes['length'] = super_pipes.length
-
-    print('list of aggregated pipes: ', aggregated_pipes)
-    print('list of aggregated forks: ', aggregated_forks)
-
-    # plotten
-    _, ax = plt.subplots()
-    super_pipes.plot(ax=ax, color='red')
-    # consumers.plot(ax=ax, color='green')
-    producers.plot(ax=ax, color='blue')
-    super_forks.plot(ax=ax, color='grey')
-    plt.title('Geometry after aggregation of pipes')
-    plt.show()
-
-    # Exportieren als geojson
-    super_forks.to_file('super_forks.geojson', driver='GeoJSON')
-    super_pipes.to_file('super_pipes.geojson', driver='GeoJSON')
-
-
-    # return
-    return {
-        'super_forks': super_forks,
-        'super_consumers': consumers, # not yet defined
-        'super_producers': producers, # not yet defined
-        'super_pipes': super_pipes,
-    }
+        forks = self.components["forks"]
+        pipes = self.components["pipes"]
+        consumers = self.components["consumers"]
+        producers = self.components["producers"]
+        self.aggregatednetwork = aggregation(forks, pipes, consumers, producers)
